feat: moved the Stats Subsystem to its own file - minor cleanup
